Separate event execution code into distinct file
