+++ src/Engine/Events/CMakeLists.txt
@@ -3,11 +3,14 @@ cmake_minimum_required(VERSION 3.20.4 FATAL_ERROR)
 set(ENGINE_EVENTS_SOURCES
         EventIR.cpp
         EventMap.cpp
+        EventInterpreter.cpp
         Processor.cpp)
 
 set(ENGINE_EVENTS_HEADERS
         EventIR.h
         EventMap.h
+        EventInterpreter.h
+        EventEnums.h
         Processor.h)
 
 add_library(engine_events STATIC ${ENGINE_EVENTS_SOURCES} ${ENGINE_EVENTS_HEADERS})
